+++ packages/types/database.types.ts
@@ -6760,6 +6760,7 @@ export type Database = {
           eft_enabled: boolean
           eft_reference_hint: string | null
           eft_swift_code: string | null
+          founding_offers_open: boolean
           id: boolean
           paypal_client_id: string | null
           paypal_enabled: boolean
@@ -6783,6 +6784,7 @@ export type Database = {
           eft_enabled?: boolean
           eft_reference_hint?: string | null
           eft_swift_code?: string | null
+          founding_offers_open?: boolean
           id?: boolean
           paypal_client_id?: string | null
           paypal_enabled?: boolean
@@ -6806,6 +6808,7 @@ export type Database = {
           eft_enabled?: boolean
           eft_reference_hint?: string | null
           eft_swift_code?: string | null
+          founding_offers_open?: boolean
           id?: boolean
           paypal_client_id?: string | null
           paypal_enabled?: boolean